refactor: remove navigation module move : mainActivity to ui module
