+++ app/build.gradle.kts
@@ -86,7 +86,6 @@ dependencies {
     implementation(project(":presentation:safeImageViewer"))
     implementation(project(":presentation:designSystem"))
     implementation(project(":presentation:viewModel"))
-    implementation(project(":presentation:navigation"))
     implementation(project(":domain:usecase"))
     implementation(project(":data:repository"))
     implementation(project(":presentation:safeImageViewer"))
