+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.Aflami"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".screens.mainactivity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
