+++ settings.gradle.kts
@@ -31,4 +31,3 @@ include(":data:local")
 include(":data:remote")
 include(":presentation:safeImageViewer")
 include(":data:repository")
-include(":presentation:navigation")
